backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
django
